Fix #10 - Update docs and rename package to eksupgrade
