+++ eksupgrade/src/S3Files/cluster_roles.json
