+++ eksupgrade/src/S3Files/kube-proxy-configmap.json
