+++ eksupgrade/src/S3Files/kube-proxy.json
