+++ eksupgrade/src/S3Files/vpc-cni.json
