+++ eksupgrade/src/eks_get_image_type.py
