+++ eksupgrade/src/self_managed.py
